fix bug

+++ src/lib.rs
@@ -1,9 +1,9 @@
 pub trait StringExtend {
-    fn substr(&self, start_pos: i32,end_pos:i32) -> String;
+    fn substr(&self, start_pos: usize, end_pos:usize) -> String;
 }
 
 impl StringExtend for str {
-    fn substr(&self, start_pos: i32,end_pos:i32) -> String {
+    fn substr(&self, start_pos: usize, end_pos:usize) -> String {
         let slice = &self[start_pos..end_pos];
         slice.to_string()
     }
